moved my visual studio project and merged it with the existing files so those of us using visual studio can use visual studio, hopefully those of us using visual code or emacs have no problems.
